refactor: Id 도메인 제거
Customer Id 타입을 Integer에서 Long으로 변경

+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import woowacourse.shoppingcart.entity.CustomerEntity;
 
 public interface CustomerDao {
-    int save(CustomerEntity customerEntity);
+    long save(CustomerEntity customerEntity);
 
     Optional<CustomerEntity> findById(long id);
